restructuring, added static folder
