+++ packages/core/__tests__/utils/getNavigationRouteForLoginScreen.spec.ts
@@ -1,7 +1,7 @@
-import { getNavigationRouteForLoginScreen } from "../../utils/getNavigationRouteForLoginScreen";
+import { getNavigationRouteForLoginScreen } from "../../src/utils/getNavigationRouteForLoginScreen";
 
 describe("getNavigationRouteForLoginScreen", () => {
-    it("should always return 'ReleaseNotes' when demo mode is set to true", () => {
+    it("should always return 'WhatsNew' when demo mode is set to true", () => {
         const route1 = getNavigationRouteForLoginScreen({
             appVersion: "1.0.0",
             lastDisplayedReleaseNotesVersion: "1.0.0",
@@ -18,12 +18,12 @@ describe("getNavigationRouteForLoginScreen", () => {
             isDemoMode: true,
         });
 
-        expect(route1).toBe("ReleaseNotes");
-        expect(route2).toBe("ReleaseNotes");
-        expect(route3).toBe("ReleaseNotes");
+        expect(route1).toBe("WhatsNew");
+        expect(route2).toBe("WhatsNew");
+        expect(route3).toBe("WhatsNew");
     });
 
-    it("Should return 'Root' if lastDisplayedReleaseNotesVersion is equal or greater than appVersion. Otherwise is should return 'ReleaseNotes'", () => {
+    it("Should return 'Root' if lastDisplayedReleaseNotesVersion is equal or greater than appVersion. Otherwise is should return 'WhatsNew'", () => {
         // Protip: Do not create nested for-loops to test all single-digit versions from '0.0.0' to '9.9.9'.
         // Doing so will result in a test that takes approx 71 seconds to run on an intel Mac.
         // Here is a simplified version
@@ -44,7 +44,7 @@ describe("getNavigationRouteForLoginScreen", () => {
             appVersion: "0.0.10",
             lastDisplayedReleaseNotesVersion: "0.0.0",
         });
-        expect(val3).toBe("ReleaseNotes");
+        expect(val3).toBe("WhatsNew");
 
         const val4 = getNavigationRouteForLoginScreen({
             appVersion: "0.0.10",
@@ -56,7 +56,7 @@ describe("getNavigationRouteForLoginScreen", () => {
             appVersion: "0.10.0",
             lastDisplayedReleaseNotesVersion: "0.0.10",
         });
-        expect(val5).toBe("ReleaseNotes");
+        expect(val5).toBe("WhatsNew");
 
         const val6 = getNavigationRouteForLoginScreen({
             appVersion: "0.10.10",
@@ -68,12 +68,12 @@ describe("getNavigationRouteForLoginScreen", () => {
             appVersion: "10.0.0",
             lastDisplayedReleaseNotesVersion: "0.10.10",
         });
-        expect(val7).toBe("ReleaseNotes");
+        expect(val7).toBe("WhatsNew");
 
         const val8 = getNavigationRouteForLoginScreen({
             appVersion: "0.0.0",
             lastDisplayedReleaseNotesVersion: "-1",
         });
-        expect(val8).toBe("ReleaseNotes");
+        expect(val8).toBe("WhatsNew");
     });
 });
