--- packages/core/src/components/screens/release-notes/mock-data.json
@@ -1,7 +1,0 @@
-{
-  "app": "Expense & Pay",
-  "version": "1.1.0",
-  "modified": "2022-12-19T07:04:55+0000",
-  "releaseNote": "- Minor bug fixes",
-  "releaseDate": "2022-12-19T07:04:55+0000"
-}
