+++ packages/core/src/static/README.md
@@ -1,0 +1,4 @@
+Do you have any static files that needs to be copied to the `dist` folder when you build? Then look
+no further! This is the correct spot to put them!
+
+`tsup` is configured to copy the whole `static` folder into `dist`.
