+++ packages/core/src/static/mock-data/whats-new.json
@@ -1,0 +1,7 @@
+{
+    "app": "App name",
+    "version": "1.1.0",
+    "modified": "2022-12-19T07:04:55+0000",
+    "releaseNote": "- Minor bug fixes",
+    "releaseDate": "2022-12-19T07:04:55+0000"
+}
