+++ packages/core/src/utils/getNavigationRouteForLoginScreen.ts
@@ -3,14 +3,14 @@ export type GetNavigationRouteForLoginScreenOptions = {
     lastDisplayedReleaseNotesVersion: string;
     isDemoMode?: boolean;
 };
-export type GetNavigationRouteForLoginScreenReturnType = "ReleaseNotes" | "Root";
+export type GetNavigationRouteForLoginScreenReturnType = "WhatsNew" | "Root";
 export const getNavigationRouteForLoginScreen = ({
     appVersion,
     lastDisplayedReleaseNotesVersion,
     isDemoMode,
 }: GetNavigationRouteForLoginScreenOptions): GetNavigationRouteForLoginScreenReturnType => {
-    if (isDemoMode) return "ReleaseNotes";
-    if (lastDisplayedReleaseNotesVersion < appVersion) return "ReleaseNotes";
+    if (isDemoMode) return "WhatsNew";
+    if (lastDisplayedReleaseNotesVersion < appVersion) return "WhatsNew";
     return "Root";
 };
