+++ packages/core/tsconfig.json
@@ -8,9 +8,9 @@
         "noPropertyAccessFromIndexSignature": true,
         "noImplicitReturns": true,
         "noFallthroughCasesInSwitch": true,
-        //"exactOptionalPropertyTypes": true,
+        "exactOptionalPropertyTypes": true,
         "outDir": "./dist"
     },
-    "include": ["./src"],
-    "exclude": ["./src/__tests__", "dist"]
+    "include": ["./src", "__tests__", "__mocks__"],
+    "exclude": ["__tests__", "dist"]
 }
